web-administrator/back-end/requirements.txt
pydantic~=1.8.2
websockets~=9.1
websocket-client
mysql-connector-python
uvicorn
fastapi==0.70.1
starlette==0.16.0
requests~=2.25.1
aiofiles
Jinja2==3.0.3
passlib[bcrypt]
python-jose[cryptography]
numpy
Pillow
pandas
python-multipart
pyyaml==5.4.1
pandas
python-dotenv==0.21.1
python-multipart
pyyaml==5.4.1
pydruid==0.6.5